requirements.txt
# Para el manejo de variables de entorno (.env)
python-dotenv==1.0.1

# Para la conexión con la base de datos PostgreSQL
psycopg2-binary==2.9.9

# Para la clonación y manejo de repositorios Git
GitPython==3.1.43


#pip install -r requirements.txt